requirements.txt
# Core Dependencies
python-swisseph>=2.10.3
pandas>=2.0.0
fastapi>=0.104.0
pydantic>=2.0.0
uvicorn>=0.24.0

# LangGraph for orchestration
langgraph>=0.0.20
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# Vector storage and embeddings
chromadb>=0.4.0
langchain-openai>=0.0.5
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.0

# PDF processing
pypdf>=3.17.0
pypdfium2>=4.26.0

# Utilities
python-dateutil>=2.8.2
tiktoken>=0.5.0
